+++ CHANGELOG.md
@@ -11,6 +11,7 @@ SPDX-License-Identifier: HEESIL-1.0
 Added:
 - links from profile page to contact page
 - Keywords
+- field to set the country of origin for devices & platforms to support export control
 
 Changed:
 - Devices can now also be mounted on other devices